forgot to show you this earlier
here's the thing in case for reference for whatever, it has changed a lot, but the ai is actually pretty decent and wins most of the time as first player, but it won't even try as second player. which I feel is fine because if someone plays decent as first you will never win as second.